lint

+++ src/maxdiffusion/models/wan/transformers/transformer_wan.py
@@ -315,9 +315,7 @@ class WanTransformerBlock(nnx.Module):
     )
 
     # 1. Self-attention
-    norm_hidden_states = (self.norm1(hidden_states) * (1 + scale_msa) + shift_msa).astype(
-        hidden_states.dtype
-    )
+    norm_hidden_states = (self.norm1(hidden_states) * (1 + scale_msa) + shift_msa).astype(hidden_states.dtype)
     attn_output = self.attn1(
         hidden_states=norm_hidden_states, encoder_hidden_states=norm_hidden_states, rotary_emb=rotary_emb
     )
@@ -329,13 +327,9 @@ class WanTransformerBlock(nnx.Module):
     hidden_states = hidden_states + attn_output
 
     # 3. Feed-forward
-    norm_hidden_states = (self.norm3(hidden_states) * (1 + c_scale_msa) + c_shift_msa).astype(
-        hidden_states.dtype
-    )
+    norm_hidden_states = (self.norm3(hidden_states) * (1 + c_scale_msa) + c_shift_msa).astype(hidden_states.dtype)
     ff_output = self.ffn(norm_hidden_states)
-    hidden_states = (hidden_states + ff_output * c_gate_msa).astype(
-        hidden_states.dtype
-    )
+    hidden_states = (hidden_states + ff_output * c_gate_msa).astype(hidden_states.dtype)
     return hidden_states
